AgriAssist/backend/requirements.txt
fastapi
uvicorn
google-generativeai
sentence-transformers
faiss-cpu
numpy
pydantic
